refactor(skills): share Gateway skill status types
Keep the emitted producer and UI JavaScript unchanged while giving the
skills.status payload one type-only owner. Preserve existing import names.

Proof: UI typecheck and 16 status tests pass; emitted JavaScript is identical.
The broad Gateway run passed 5443 tests but the unrelated cache-retention
child hit its existing 15s deadline twice. Preserve as draft pending gates.
