+++ src/skills/discovery/status.ts
@@ -3,15 +3,12 @@ import path from "node:path";
 import type { OpenClawConfig } from "../../config/types.openclaw.js";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
 import { evaluateEntryRequirementsForCurrentPlatform } from "../../shared/entry-status.js";
-import type { RequirementConfigCheck, Requirements } from "../../shared/requirements.js";
 import { CONFIG_DIR } from "../../utils.js";
 import {
   readClawHubSkillsLockfileStatusSync,
   resolveClawHubSkillStatusLinkSync,
   resolveLocalSkillCardStatusSync,
-  type ClawHubSkillStatusLink,
   type ClawHubSkillsLockfileStatusRead,
-  type LocalSkillCardStatus,
 } from "../lifecycle/clawhub.js";
 import { resolveBundledSkillsDir } from "../loading/bundled-dir.js";
 import {
@@ -37,58 +34,11 @@ import {
   normalizeSkillIndexName,
   type SkillIndexEntry,
 } from "./skill-index.js";
+import type { SkillInstallOption, SkillStatusEntry, SkillStatusReport } from "./status.types.js";
+export type { SkillStatusEntry, SkillStatusReport } from "./status.types.js";
 
 const skillsLogger = createSubsystemLogger("skills");
 let hasWarnedMissingBundledDir = false;
-
-type SkillInstallOption = {
-  id: string;
-  kind: SkillInstallSpec["kind"];
-  label: string;
-  bins: string[];
-};
-
-export type SkillStatusEntry = {
-  name: string;
-  description: string;
-  source: string;
-  bundled: boolean;
-  filePath: string;
-  baseDir: string;
-  skillKey: string;
-  primaryEnv?: string;
-  emoji?: string;
-  homepage?: string;
-  always: boolean;
-  disabled: boolean;
-  blockedByAllowlist: boolean;
-  blockedByAgentFilter: boolean;
-  eligible: boolean;
-  /**
-   * True when the skill declares an OS requirement that does not include the
-   * current platform (e.g. a macOS-only skill on Linux/Windows). Such skills are
-   * inapplicable by design rather than broken installs, so callers can surface
-   * them separately from genuine "missing requirements".
-   */
-  platformIncompatible: boolean;
-  modelVisible: boolean;
-  userInvocable: boolean;
-  commandVisible: boolean;
-  requirements: Requirements;
-  missing: Requirements;
-  configChecks: RequirementConfigCheck[];
-  install: SkillInstallOption[];
-  clawhub?: ClawHubSkillStatusLink;
-  skillCard?: LocalSkillCardStatus;
-};
-
-export type SkillStatusReport = {
-  workspaceDir: string;
-  managedSkillsDir: string;
-  agentId?: string;
-  agentSkillFilter?: string[];
-  skills: SkillStatusEntry[];
-};
 
 export function resolveSkillStatusEntry<T extends Pick<SkillStatusEntry, "name" | "skillKey">>(
   skills: readonly T[],
